Fixing warnings.
